# hw2/api/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: CRUD API for Marketplace
servers:
  - url: http://localhost:8080/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        stock:
          type: integer
        category:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
        seller_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - seller_id
        - created_at
        - updated_at
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
        price:
          type: number
          format: float
          minimum: 0.01
          exclusiveMinimum: true
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
      required:
        - name
        - price
        - stock
        - category
        - status
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
        price:
          type: number
          format: float
          minimum: 0.01
          exclusiveMinimum: true
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
    ProductListResponse:
      type: object
      properties:
        totalElements:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
      required:
        - totalElements
        - pageNumber
        - pageSize
        - products
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error_code
        - message
...
    PromoCodeCreate:
      type: object
      properties:
        code:
          type: string
          pattern: ^[A-Z0-9_]{4,20}$
        discount_type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        discount_value:
          type: number
          format: float
        min_order_amount:
          type: number
          format: float
        max_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
      required:
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - valid_from
        - valid_until
    PromoCodeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          type: string
        discount_value:
          type: number
          format: float
        min_order_amount:
          type: number
          format: float
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
      required:
        - id
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - current_uses
        - valid_from
        - valid_until
        - active
paths:
...
  /products:
    post:
      security:
        - bearerAuth: []
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access Denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, ARCHIVED]
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Product List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{id}:
    get:
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      security:
        - bearerAuth: []
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    delete:
      security:
        - bearerAuth: []
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted (Archived)
...
  /promo-codes:
    post:
      security:
        - bearerAuth: []
      operationId: createPromoCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'